cleanup lpa
